@@ -1,17 +1,15 @@
 name: find_dropdown
 description: Simple and robust FindDropdown with item search feature, making it possible to use an offline item list or filtering URL for easy customization.
-version: 0.2.2
-author: F-Team <david@flutterando.com.br>
+version: 1.0.0
 homepage: https://flutterando.com.br/
 git: https://github.com/davidsdearaujo/find_dropdown
 repository: https://github.com/davidsdearaujo/find_dropdown
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  rxdart: ">=0.23.0 <0.25.0"
-  select_dialog: ^1.2.1
+  select_dialog: ^2.0.0
   flutter:
     sdk: flutter
 
