@@ -2,14 +2,13 @@
 
 class UserModel {
   final String id;
-  final DateTime createdAt;
+  final DateTime? createdAt;
   final String name;
   final String avatar;
 
-  UserModel({this.id, this.createdAt, this.name, this.avatar});
+  UserModel({required this.id, required this.createdAt, required this.name, required this.avatar});
 
   factory UserModel.fromJson(Map<String, dynamic> json) {
-    if (json == null) return null;
     return UserModel(
       id: json["id"],
       createdAt: json["createdAt"] == null ? null : DateTime.parse(json["createdAt"]),
@@ -19,7 +18,6 @@ class UserModel {
   }
 
   static List<UserModel> fromJsonList(List list) {
-    if (list == null) return null;
     return list.map((item) => UserModel.fromJson(item)).toList();
   }
 
